Using floats instead of doubles and hacked in animation

+++ src/main/java/as2388/scala/raytracer/IntersectionData.scala
@@ -2,4 +2,4 @@ package as2388.scala.raytracer
 
 import as2388.scala.raytracer.shapes.Shape
 
-class IntersectionData (val distance: Double, val intersectionPoint: Point, val normal: Vector, val shape: Shape)
+class IntersectionData (val distance: Float, val intersectionPoint: Point, val normal: Vector, val shape: Shape)
