+++ src/main/java/as2388/scala/raytracer/Singularity.scala
@@ -1,3 +1,3 @@
 package as2388.scala.raytracer
 
-class Singularity(val location: Point, val strength: Double)
+class Singularity(val location: Point, val strength: Float)
